requirements.txt
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
notebook
langchain-tavily
wikipedia
trustcall
langgraph-cli[inmem]
python-dotenv
upstash-redis
pypdf
faiss-cpu
langchain-text-splitters
redis
psycopg2-binary

